Update README.md
Flipkart-Frontend – E-commerce UI in React 
This is the frontend of a full-stack Flipkart Clone project built using React.js, designed to offer a smooth and responsive e-commerce experience. The application interfaces with a Node.js/Express backend and MongoDB database to deliver real-time product display, cart functionality, and user authentication.